Programmatically Updating all Broken Links
We have written a program that leverages Electric Capital's index
to review all histories of Web3 / crypto (or any git repo for that
matter) for complexity.

The first step in this process is to clone that repository. We noticed
that a huge subset of the repositories in this index are no longer
accessible.

This is a draft PR to share this information back with those who depend
on it. It would be great to have a conversation to see what else we
could do in terms of base-level foundation data in web3/crypto.

Please let me know if you'd like other things done programmatically at
scale.

I'm reachable at adam@foliofficient.com