cleaning up ronin & axie entries
* correcting ronin github org and connection to axie sub-eco
* removing organization repos, as per README.md

"It's important to note also that you do not need to add all the repos within your GitHub organizations to the .toml file as individual repos, because the system automatically fetches all repos within the organization. [...] This is unnecessary. It adds clutter and makes it harder for reviewers to approve your PR."