Add web3:// protocol sub-ecosystem to Ethereum.
The GitHub Web3 Protocol organization is home to libraries and
implementations of the
[IANA-standardized `web3://` URI scheme](https://www.iana.org/assignments/uri-schemes/uri-schemes.xhtml),
defined in Ethereum's [ERC-6860](https://eips.ethereum.org/EIPS/eip-6860).

This protocol puts entire website logic onchain and essentially turns an EVM
chain into an execution environment for web server code.

The list of listed repositories includes implementations of ERC-6860 client
libraries, as well as a web browser extension and a Chromium fork that
implements the `web3://` URI scheme.